refactor(core): move lifecycle state owners out of core
Move provider-neutral session, turn, confirmation, question, cancellation, and background exec-output state into execution-layer owner crates.

Keep bitfun-core compatibility facades and add boundary guards so migrated owners do not regress into core.
