+++ src/crates/execution/agent-runtime/AGENTS.md
@@ -15,9 +15,10 @@ port-backed `AgentRuntime` facade that can be built and tested without
   must not create concrete managers, app state, filesystem, terminal, MCP,
   remote, or AI clients.
 - Keep concrete scheduler/session lifecycle execution, session metadata IO,
-  event emitter wiring, permission UI / channel waits, and product `Tool`
-  adapter execution in `bitfun-core` until a reviewed owner migration proves
-  behavior equivalence.
+  event emitter wiring, permission UI presentation, and product `Tool` adapter
+  execution in `bitfun-core` until a reviewed owner migration proves behavior
+  equivalence. Provider-neutral confirmation/user-question wait channel state
+  may live here.
 - Prefer pure facts and decisions first: queue policy, background delivery,
   dialog-turn queue state, active-turn facts, cancellation routing and
   suppression state, background running-turn injection construction, steering action
@@ -26,21 +27,24 @@ port-backed `AgentRuntime` facade that can be built and tested without
   registry visibility/availability, custom subagent schema/default decisions,
   builtin agent definition catalog, thread-goal metadata / event payload /
   token usage / scheduler delivery plans, thread-goal tool wire contracts,
-  user-question validation/result contracts, SessionControl input/cancel-route/result contracts, DeepReview
+  user-question validation/result/channel contracts, SessionControl input/cancel-route/result contracts, DeepReview
   policy/manifest/budget/queue/report/cache/shared-context/task-execution
   shaping decisions, DeepResearch citation renumbering,
   custom subagent markdown front-matter IO, custom subagent discovery/loading,
   post-call hook routing/executor orchestration,
-  tool confirmation planning/failure/wait-result mapping, light checkpoint
-  summary policy, round-boundary yield/injection state, turn-outcome
+  tool confirmation planning/failure/wait-result/channel mapping, light checkpoint
+  summary policy, dialog-turn cancellation token state,
+  round-boundary yield/injection state, turn-outcome
   queue decisions, registry source/profile facts, prompt-loop user-context
   policy, prompt listing reminder ordering, prompt-cache policy/identity/store,
-  finish-reason labels, session-state event labels, and turn-outcome event facts.
+  turn skill/agent snapshot state, file-read session state, session evidence
+  ledger projection, finish-reason labels, session-state event labels, and
+  turn-outcome event facts.
 - Keep concrete prompt assembly, workspace context IO, prompt-cache persistence
   wiring, dynamic environment collection, concrete hook side effects,
   DeepReview task launch/provider wait/report persistence, DeepResearch
-  storage IO/post-turn hook, concrete product tool execution, and channel/state
-  mutation outside this crate until a reviewed migration proves behavior
+  storage IO/post-turn hook and concrete product tool execution outside this
+  crate until a reviewed migration proves behavior
   equivalence.
 - Add focused tests before moving any runtime decision into this crate.
