+++ src/crates/execution/agent-runtime/src/lib.rs
@@ -9,6 +9,8 @@ pub mod custom_subagent;
 pub mod deep_research;
 pub mod deep_review;
 pub mod events;
+pub mod evidence_ledger;
+pub mod file_read_state;
 pub mod post_call_hooks;
 pub mod prompt;
 pub mod prompt_cache;
@@ -16,7 +18,9 @@ pub mod runtime;
 pub mod scheduled_job;
 pub mod scheduler;
 pub mod session_control;
+pub mod skill_agent_snapshot;
 pub mod thread_goal;
 pub mod thread_goal_tools;
 pub mod tool_confirmation;
+pub mod turn_cancellation;
 pub mod user_questions;
