+++ src/crates/execution/tool-execution/Cargo.toml
@@ -4,10 +4,14 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
+dashmap = { workspace = true }
 globset = { workspace = true }
 dunce = { workspace = true }
 grep-regex = { workspace = true }
 grep-searcher = { workspace = true }
 ignore = { workspace = true }
 log = { workspace = true }
+serde = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
 vte = { workspace = true, features = ["ansi"] }
